test(transfer): add command integration tests
